# 06-API-NodeJS/02-swagger/01-blog-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "Blog API - Users, Posts, Comments" 
  version: "1.0.0"
  title: "Swagger Blog"
  contact:
    email: "contact@afpa.fr"
host: "localhost:3000"
...
schemes:
- "https"
- "http"
paths:
...
  /posts:
    get:
      tags:
      - "Posts"
      summary: "Get all posts"
      description: "Get all posts"
      operationId: "getPosts"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Get all posts"
        required: true
        schema:
          $ref: "#/definitions/Posts"
      responses:
        "405":
          description: "Invalid input"
           
    post:
      tags:
      - "Posts"
      summary: "Create a new post"
      description: "Create a new post"
      operationId: "createPost"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Create a new post"
        required: true
        schema:
          $ref: "#/definitions/Posts"
      responses:
        "405":
          description: "Invalid input"
  /posts/:id:
    get:
      tags:
      - "Posts"
      summary: "Get a user with his id"
      description: "Get a user with his id"
      operationId: "getPostById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "user_id"
        in: "path"
        description: "User id"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "User not found"
        
    delete:
      tags:
      - "Posts"
      summary: "Delete a user by id"
      description: "Delete a user by id"
      operationId: "deletePost"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "postId"
        in: "path"
        description: "Post id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Post id not found"
       
    put:
      tags:
      - "Posts"
      summary: "Update an existing post"
      description: "Update an existing post"
      operationId: "updatePost"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Update an existing post"
        required: true
        schema:
          $ref: "#/definitions/Posts"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Post id not found"
        "405":
          description: "Validation exception"
...
definitions:
...
  Posts:
    type: "object"
    properties:
      user_id: 
        type: "integer"
        format: "int64"
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      body: 
        type: "string"
    xml:
      name: "Posts"
...